feat(import existing): trigger import of attached resources of a note into the note
A new feature arrived to import existing note attachment (files) to a existing note.
The attachment must be connected to the note, then the new feature is able to import and use the same logic like for new stuff. It will import the resource of a note into the note.
There are two options
Override note body
Append to note body